refactor: move MegaMenu to separate file
